moved helper functions to util.py file
